feat(email): PR Fixes
